+++ react-components/src/architecture/concrete/primitives/line/LineCreator.ts
@@ -3,12 +3,12 @@
  */
 
 import { Plane, type Ray, Vector3 } from 'three';
-import { PrimitiveType } from './PrimitiveType';
-import { BaseCreator } from '../../base/domainObjectsHelpers/BaseCreator';
-import { copy } from '../../base/utilities/extensions/arrayExtensions';
-import { Changes } from '../../base/domainObjectsHelpers/Changes';
-import { type DomainObject } from '../../base/domainObjects/DomainObject';
-import { FocusType } from '../../base/domainObjectsHelpers/FocusType';
+import { PrimitiveType } from '../PrimitiveType';
+import { BaseCreator } from '../../../base/domainObjectsHelpers/BaseCreator';
+import { copy } from '../../../base/utilities/extensions/arrayExtensions';
+import { Changes } from '../../../base/domainObjectsHelpers/Changes';
+import { type DomainObject } from '../../../base/domainObjects/DomainObject';
+import { FocusType } from '../../../base/domainObjectsHelpers/FocusType';
 import { type LineDomainObject } from './LineDomainObject';
 
 /**
